Integrated into queryBuilder
