+++ tests/integration/batch.test.ts
@@ -1,6 +1,5 @@
 import makeServer from './drivers/default/server';
 import Post from '../stubs/models/post';
-import { Batch } from '../../src/batch';
 import { Orion } from '../../src/orion';
 
 let server: any;
@@ -22,7 +21,7 @@ describe('Batch tests', () => {
 		posts[0].$attributes.title = "First";
 		posts[1].$attributes.title = "Second";
 
-		const res = await Batch.store(posts);
+		const res = await Post.$query().batchStore(posts);
 
 		expect(server.schema.posts.all()).toHaveLength(2);
 		expect(server.schema.posts.find('1').attrs.title).toBe("First")
@@ -43,12 +42,12 @@ describe('Batch tests', () => {
 		posts[1].$attributes.title = "Second";
 		posts[2].$attributes.title = "Third";
 
-		let res = await Batch.store(posts);
+		let res = await Post.$query().batchStore(posts);
 
 		res[0].$attributes.title = "NewFirst";
 		res[1].$attributes.title = "NewSecond";
 
-		res = await Batch.update([res[0],res[1]]);
+		res = await Post.$query().batchUpdate([res[0],res[1]]);
 
 		expect(res).toHaveLength(2);
 		expect(server.schema.posts.find('1').attrs.title).toBe("NewFirst")
@@ -69,16 +68,15 @@ describe('Batch tests', () => {
 		posts[1].$attributes.title = "Second";
 		posts[2].$attributes.title = "Third";
 
-		let res = await Batch.store(posts);
+		let res = await Post.$query().batchStore(posts);
 
-		let ModelDelete = await Batch.delete([res[1]]);
-		let idDelete = await Batch.delete([3], new Post);
+		let ModelDelete = await Post.$query().batchDelete([res[1].$getKey(), res[2].$getKey()]);
 
 		expect(server.schema.posts.find('1').attrs.deleted_at).toBeUndefined();
 		expect(server.schema.posts.find('2').attrs.deleted_at).toBeDefined();
 		expect(server.schema.posts.find('3').attrs.deleted_at).toBeDefined();
 		expect(server.schema.posts.find('2').attrs.title).toEqual(ModelDelete[0].$attributes.title)
-		expect(server.schema.posts.find('3').attrs.title).toEqual(idDelete[0].$attributes.title)
+		expect(server.schema.posts.find('3').attrs.title).toEqual(ModelDelete[1].$attributes.title)
 
 
 	});
@@ -93,13 +91,12 @@ describe('Batch tests', () => {
 		posts[1].$attributes.title = "Second";
 		posts[2].$attributes.title = "Third";
 
-		let res = await Batch.store(posts);
+		let res = await Post.$query().batchStore(posts);
 
 		// delete ID 2 & 3
-		let ModelDelete = await Batch.delete([res[1]]);
-		let idDelete = await Batch.delete([3], new Post);
+		let ModelDelete = await Post.$query().batchDelete([res[1].$getKey(), res[2].$getKey()]);
 
-		res = await Batch.restore([...ModelDelete, ...idDelete]);
+		res = await Post.$query().batchRestore(ModelDelete.map(x => x.$getKey()));
 
 		expect(server.schema.posts.find('1').attrs.deleted_at).toBeFalsy();
 		expect(server.schema.posts.find('2').attrs.deleted_at).toBeFalsy();
